[fix] 유닛 소환 데이터 구조 변경

+++ Assets/01.Scripts/02.Entity/Unit/Summon/SummonModel.cs
@@ -7,7 +7,7 @@ public class SummonModel : BaseModel
     public string rcode;
     public int level;
     public int cost;
-    public float[] summonRate;
+    public int[] summonRate;
 
     public SummonModel(SummonTableSO summonTableSO)
     {
